Split settings from randomizer control in UI
